+++ src/components/App.js
@@ -1,13 +1,12 @@
 import React, {Component} from 'react';
 import '../styles/App.css';
-import Landing from "../pages/Landing";
+import Landing from "./Landing";
 import Scroller from "./Scroller";
-import Members from "../pages/Members";
-import Team from "../pages/Team";
-import First from "../pages/First";
-import Contact from "../pages/Contact";
-import Sponsors from "../pages/Sponsors";
-import Info from "../pages/Info";
+import Members from "./Members";
+import Team from "./Team";
+import First from "./First";
+import Contact from "./Contact";
+import Sponsors from "./Sponsors";
 
 class App extends Component {
   static scrollTo(id) {
@@ -21,10 +20,8 @@ class App extends Component {
       <Scroller pages={[
         {
           name: "Home", id: "landing", component: <Landing/>
-        },{
-          name: "Info", id: "info", hide: true, component: <Info/>
         }, {
-          name: "Meet the Gearheads", id: "members", component: <Members/>
+          name: "Members the Gearheads", id: "members", component: <Members/>
         }, {
           name: "Sponsors", id: "sponsors", component: <Sponsors/>
         }, {
