+++ src/styles/all-members.css
@@ -29,7 +29,6 @@
 
     height: calc(100% - 64px);
     width: calc(100% - 64px);
-    max-width: 1440px;
 
     background: white;
     color: black;
@@ -37,7 +36,6 @@
     padding: 32px 32px 0;
 
     overflow: auto;
-    border-radius: 32px;
 }
 
 .all-members .button {
@@ -51,7 +49,7 @@
     text-align: center;
 }
 
-.all-members h3, .all-members p {
+.all-members h3, p {
     text-align: center;
 }
 
@@ -73,8 +71,8 @@
 
 .all-members-container button {
     position: absolute;
-    top: 16px;
-    right: 16px;
+    top: 0;
+    right: 0;
 
     border: 0;
     background: 0;
