+++ src/styles/first.css
@@ -17,8 +17,11 @@
     margin: 0 auto 15px;
 }
 
-#first h1 {
+#first h3 {
     font-size: 36pt;
+
+    font-family: 'IBM Plex Sans', sans-serif;
+    font-weight: 100;
 }
 
 #first p {
@@ -42,7 +45,7 @@
         width: 30%;
     }
 
-    #first h1 {
+    #first h3 {
         font-size: 24pt;
     }
