--- src/styles/info.css
@@ -1,112 +1,0 @@
-#info-page {
-    padding-left: 0;
-    background-image: url('../images/info.JPG');
-
-    text-align: left;
-}
-
-#info-page::after {
-    opacity: .8;
-}
-
-#info {
-    min-height: 100vh;
-    display: flex;
-}
-
-#info h1 {
-    font-size: 4vw;
-    margin-bottom: 0;
-}
-
-#info h3 {
-    margin-top: 0;
-    font-size: 2vw;
-}
-
-#info p {
-    font-size: 1.2vw;
-}
-
-#info img {
-    width: 100%;
-    max-width: 480px;
-}
-
-#info .half-white {
-    position: relative;
-    min-height: 100vh;
-    width: 35%;
-    background: rgba(255, 255, 255, 0.9);
-    color: black;
-    padding-left: 64px;
-    padding-top: 64px;
-    padding-bottom: 128px;
-}
-
-#info .half-white .info-diagonal {
-    position: absolute;
-    width: 20vw;
-    height: 100%;
-    top: 0;
-    right: -20vw;
-    opacity: 0.9;
-}
-
-#info .half-white .info-diagonal img {
-    height: 100%;
-    width: 100%;
-    max-width: none;
-}
-
-#info .half {
-    position: relative;
-    min-height: 100vh;
-    width: 40%;
-    color: white;
-    padding-right: 48px;
-    padding-top: 128px;
-    padding-bottom: 64px;
-    margin-left: auto;
-}
-
-@media (max-width: 1440px) {
-    #info p {
-        font-size: 18px;
-    }
-}
-
-@media (max-width: 768px) {
-    #info h1 {
-        font-size: 48px;
-    }
-
-    #info h3 {
-        font-size: 24px;
-    }
-
-    #info .half-white {
-        min-height: 30vh;
-        float: none;
-        width: 100%;
-        padding: 32px;
-    }
-
-    #info .half-white .info-diagonal {
-        position: absolute;
-        width: 100vw;
-        height: 20vh;
-        top: initial;
-        left: 0;
-        bottom: -20vh;
-        opacity: 0.9;
-    }
-
-    #info .half {
-        margin-top: 20vh;
-        min-height: 30vh;
-        float: none;
-        width: 100%;
-        padding: 64px;
-    }
-}
